+++ Sources/StreamChatSwiftUI/DefaultViewFactory.swift
@@ -370,7 +370,15 @@ extension ViewFactory {
     public func makeLastInGroupHeaderView(for message: ChatMessage) -> some View {
         EmptyView()
     }
-    
+
+    public func makeMessageTranslationFooterView(
+        messageViewModel: MessageViewModel
+    ) -> some View {
+        MessageTranslationFooterView(
+            messageViewModel: messageViewModel
+        )
+    }
+
     public func makeImageAttachmentView(
         for message: ChatMessage,
         isFirst: Bool,
@@ -452,11 +460,20 @@ extension ViewFactory {
         options: MediaViewsOptions
     ) -> some View {
         GalleryView(
+            viewFactory: self,
             mediaAttachments: mediaAttachments,
             author: message.author,
             isShown: isShown,
             selected: options.selectedIndex
         )
+    }
+    
+    public func makeGalleryHeaderView(
+        title: String,
+        subtitle: String,
+        shown: Binding<Bool>
+    ) -> some View {
+        GalleryHeaderView(title: title, subtitle: subtitle, isShown: shown)
     }
     
     public func makeVideoPlayerView(
@@ -466,10 +483,19 @@ extension ViewFactory {
         options: MediaViewsOptions
     ) -> some View {
         VideoPlayerView(
+            viewFactory: self,
             attachment: attachment,
             author: message.author,
             isShown: isShown
         )
+    }
+    
+    public func makeVideoPlayerHeaderView(
+        title: String,
+        subtitle: String,
+        shown: Binding<Bool>
+    ) -> some View {
+        GalleryHeaderView(title: title, subtitle: subtitle, isShown: shown)
     }
     
     public func makeDeletedMessageView(
@@ -986,7 +1012,7 @@ extension ViewFactory {
             currentUserId: chatClient.currentUserId,
             message: message
         )
-        let showReadCount = channel.memberCount > 2
+        let showReadCount = channel.memberCount > 2 && !message.isLastActionFailed
         return MessageReadIndicatorView(
             readUsers: readUsers,
             showReadCount: showReadCount,